pyproject.toml
[tool.poetry]
name = "toi-project1"
version = "0.1.0"
description = ""
authors = ["Marek Vasko <xvasko16@vutbr.cz>"]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
numpy = "^1.22.3"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
